let agent run on windows
